@@ -1,7 +1,12 @@
 {
+  "homepage": "https://awesome-app.vercel.app",
   "name": "client",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": "16.16.0",
+    "npm": "8.11.0"
+  },
   "dependencies": {
     "@david.kucsai/react-pdf-table": "^0.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
@@ -9,6 +14,8 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
+    "@react-pdf/font": "2.2.0",
+    "@react-pdf/pdfkit": "^3.0.2",
     "@react-pdf/renderer": "^2.0.19",
     "@stripe/react-stripe-js": "^1.4.1",
     "@stripe/stripe-js": "^1.17.1",
@@ -17,31 +24,32 @@
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
     "firebase": "^8.6.8",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
-    "react-bootstrap": "^1.6.1",
+    "react-bootstrap": "^2.6.0",
     "react-datepicker": "^4.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-image-file-resizer": "^0.4.7",
     "react-loading-skeleton": "^2.2.0",
     "react-modal-image": "^2.5.0",
     "react-multi-select-component": "^4.0.6",
     "react-paginate": "^7.1.3",
+    "react-pdf": "^7.3.3",
     "react-redux": "^7.2.4",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "4.0.3",
     "react-slick": "^0.28.1",
     "redux": "^4.1.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0",
-    "slick-carousel": "^1.8.1",
-    "web-vitals": "^1.1.2"
+    "slick-carousel": "^1.8.1"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts build",
+    "start": "set PORT=3000 && react-scripts --openssl-legacy-provider start",
+    "build": "react-scripts --openssl-legacy-provider build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "build:dev": "env-cmd -f .env.development npm run build && firebase deploy -P dev",
@@ -66,6 +74,11 @@
     ]
   },
   "devDependencies": {
-    "env-cmd": "^10.1.0"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "env-cmd": "^10.1.0",
+    "web-vitals": "^3.1.0"
+  },
+  "resolutions": {
+    "@react-pdf/font": "2.2.0"
   }
 }
