@@ -1,33 +1,33 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
-import App from './App'
-import { Provider } from 'react-redux'
-import store from './store'
-import { BrowserRouter } from 'react-router-dom'
-import { transitions, positions, Provider as AlertProvider } from 'react-alert'
-import AlertTemplate from 'react-alert-template-basic'
-import reportWebVitals from './reportWebVitals'
-import './bootstrap.min.css'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import App from "./App";
+import { Provider } from "react-redux";
+import store from "./store";
+import { transitions, positions, Provider as AlertProvider } from "react-alert";
+import AlertTemplate from "react-alert-template-basic";
+import reportWebVitals from "./reportWebVitals";
+import "./bootstrap.min.css";
+import ErrorBoundary from "./components/ErroBoundary";
 const options = {
   // you can also just use 'bottom center'
   position: positions.BOTTOM_RIGHT,
   timeout: 5000,
-  offset: '30px',
+  offset: "30px",
   // you can also just use 'scale'
   transition: transitions.SCALE,
-}
+};
 
-ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter>
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
+  <ErrorBoundary>
+    <Provider store={store}>
       <AlertProvider template={AlertTemplate} {...options}>
         <App />
       </AlertProvider>
-    </BrowserRouter>
-  </Provider>,
+    </Provider>
+  </ErrorBoundary>
+);
 
-  document.getElementById('root')
-)
-
-reportWebVitals()
+reportWebVitals();
