@@ -1,4 +1,4 @@
-import axios from 'axios'
+import axios from "axios";
 import {
   APPLY_COUPON_FAIL,
   APPLY_COUPON_REQUEST,
@@ -17,7 +17,7 @@ import {
   DB_CART_CLEAR_FAIL,
   DB_CART_CLEAR_REQUEST,
   DB_CART_CLEAR_SUCCESS,
-} from '../constants/cartConstants'
+} from "../constants/cartConstants";
 
 export const addToCart =
   (slug, qty, variable = null, addonPd = null) =>
@@ -25,7 +25,7 @@ export const addToCart =
     if (slug || qty || variable || addonPd) {
       const { data } = await axios.get(
         `${process.env.REACT_APP_API}/api/product/${slug}`
-      )
+      );
 
       dispatch({
         type: CART_ADD_ITEM,
@@ -90,37 +90,40 @@ export const addToCart =
               )
               .reduce((acc, item) => acc + item.price, 0),
         },
-      })
+      });
     }
 
-    localStorage.setItem('cartItems', JSON.stringify(getState().cart.cartItems))
-  }
+    localStorage.setItem(
+      "cartItems",
+      JSON.stringify(getState().cart.cartItems)
+    );
+  };
 
 export const removeFromCart = (slug) => (dispatch, getState) => {
   dispatch({
     type: CART_REMOVE_ITEM,
     payload: slug,
-  })
-  localStorage.setItem('cartItems', JSON.stringify(getState().cart.cartItems))
-}
+  });
+  localStorage.setItem("cartItems", JSON.stringify(getState().cart.cartItems));
+};
 
 export const dbSaveCart = (cart) => async (dispatch, getState) => {
   try {
-    dispatch({ type: CART_DB_REQUEST })
+    dispatch({ type: CART_DB_REQUEST });
 
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
-    await axios.post(`${process.env.REACT_APP_API}/api/cart`, { cart }, config)
-    dispatch({ type: CART_DB_SUCCESS })
+    await axios.post(`${process.env.REACT_APP_API}/api/cart`, { cart }, config);
+    dispatch({ type: CART_DB_SUCCESS });
   } catch (error) {
     dispatch({
       type: CART_DB_FAIL,
@@ -128,29 +131,29 @@ export const dbSaveCart = (cart) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const listCart = () => async (dispatch, getState) => {
   try {
-    dispatch({ type: CART_LIST_REQUEST })
+    dispatch({ type: CART_LIST_REQUEST });
 
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
     const { data } = await axios.get(
       `${process.env.REACT_APP_API}/api/cart`,
       config
-    )
-    dispatch({ type: CART_LIST_SUCCESS, payload: data })
+    );
+    dispatch({ type: CART_LIST_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: CART_LIST_FAIL,
@@ -158,26 +161,26 @@ export const listCart = () => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const clearDbCart = () => async (dispatch, getState) => {
   try {
-    dispatch({ type: DB_CART_CLEAR_REQUEST })
+    dispatch({ type: DB_CART_CLEAR_REQUEST });
 
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
-    await axios.delete(`${process.env.REACT_APP_API}/api/cart`, config)
-    dispatch({ type: DB_CART_CLEAR_SUCCESS })
+    await axios.delete(`${process.env.REACT_APP_API}/api/cart`, config);
+    dispatch({ type: DB_CART_CLEAR_SUCCESS });
   } catch (error) {
     dispatch({
       type: DB_CART_CLEAR_FAIL,
@@ -185,31 +188,31 @@ export const clearDbCart = () => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const couponApply = (coupon) => async (dispatch, getState) => {
   try {
-    dispatch({ type: APPLY_COUPON_REQUEST })
+    dispatch({ type: APPLY_COUPON_REQUEST });
 
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
     await axios.post(
       `${process.env.REACT_APP_API}/api/cart/coupon`,
       { coupon },
       config
-    )
-    dispatch({ type: APPLY_COUPON_SUCCESS })
+    );
+    dispatch({ type: APPLY_COUPON_SUCCESS });
   } catch (error) {
     dispatch({
       type: APPLY_COUPON_FAIL,
@@ -217,30 +220,30 @@ export const couponApply = (coupon) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const couponCancel = () => async (dispatch, getState) => {
   try {
-    dispatch({ type: CANCEL_COUPON_REQUEST })
+    dispatch({ type: CANCEL_COUPON_REQUEST });
 
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
     await axios.post(
       `${process.env.REACT_APP_API}/api/cart/coupon-cancel`,
       {},
       config
-    )
-    dispatch({ type: CANCEL_COUPON_SUCCESS })
+    );
+    dispatch({ type: CANCEL_COUPON_SUCCESS });
   } catch (error) {
     dispatch({
       type: CANCEL_COUPON_FAIL,
@@ -248,31 +251,31 @@ export const couponCancel = () => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const userDbCartDelete = () => async (dispatch, getState) => {
   try {
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
 
     await axios.delete(
       `${process.env.REACT_APP_API}/api/cart/delete-user-cart`,
       config
-    )
+    );
   } catch (error) {
     alert(
       error.response && error.response.data.message
         ? error.response.data.message
         : error.message
-    )
+    );
   }
-}
+};
