@@ -1,4 +1,4 @@
-import axios from 'axios'
+import axios from "axios";
 import {
   COUNT_PRODUCTS_FAIL,
   COUNT_PRODUCTS_REQUEST,
@@ -27,28 +27,28 @@ import {
   UPLOAD_IMAGE_FAIL,
   UPLOAD_IMAGE_REQUEST,
   UPLOAD_IMAGE_SUCCESS,
-} from '../constants/productConstants'
+} from "../constants/productConstants";
 
 export const uploadFile = (file) => async (dispatch, getState) => {
   try {
-    dispatch({ type: UPLOAD_IMAGE_REQUEST })
+    dispatch({ type: UPLOAD_IMAGE_REQUEST });
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
     const { data } = await axios.post(
       `${process.env.REACT_APP_API}/api/upload`,
       { image: file },
       config
-    )
+    );
 
-    dispatch({ type: UPLOAD_IMAGE_SUCCESS, payload: data })
+    dispatch({ type: UPLOAD_IMAGE_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: UPLOAD_IMAGE_FAIL,
@@ -56,25 +56,25 @@ export const uploadFile = (file) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const removeFile = (id) => async (dispatch, getState) => {
   try {
-    dispatch({ type: REMOVE_IMAGE_REQUEST })
+    dispatch({ type: REMOVE_IMAGE_REQUEST });
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
-    await axios.post(`${process.env.REACT_APP_API}/api/remove`, { id }, config)
-    dispatch({ type: REMOVE_IMAGE_SUCCESS })
+    };
+    await axios.post(`${process.env.REACT_APP_API}/api/remove`, { id }, config);
+    dispatch({ type: REMOVE_IMAGE_SUCCESS });
   } catch (error) {
     dispatch({
       type: REMOVE_IMAGE_FAIL,
@@ -82,29 +82,29 @@ export const removeFile = (id) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const createProduct = (product) => async (dispatch, getState) => {
   try {
-    dispatch({ type: CREATE_PRODUCT_REQUEST })
+    dispatch({ type: CREATE_PRODUCT_REQUEST });
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
     await axios.post(
       `${process.env.REACT_APP_API}/api/product`,
       product,
       config
-    )
-    dispatch({ type: CREATE_PRODUCT_SUCCESS })
+    );
+    dispatch({ type: CREATE_PRODUCT_SUCCESS });
   } catch (error) {
     dispatch({
       type: CREATE_PRODUCT_FAIL,
@@ -112,18 +112,16 @@ export const createProduct = (product) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const listProduct = (search) => async (dispatch) => {
   try {
-    dispatch({ type: LIST_PRODUCT_REQUEST })
+    dispatch({ type: LIST_PRODUCT_REQUEST });
 
-    const { data } = await axios.get(
-      `${process.env.REACT_APP_API}/api/product?search=${search}`
-    )
-    dispatch({ type: LIST_PRODUCT_SUCCESS, payload: data })
+    const { data } = await axios.get(`/api/product?search=${search}`);
+    dispatch({ type: LIST_PRODUCT_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: LIST_PRODUCT_FAIL,
@@ -131,24 +129,24 @@ export const listProduct = (search) => async (dispatch) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const deleteProduct = (id) => async (dispatch, getState) => {
   try {
-    dispatch({ type: DELETE_PRODUCT_REQUEST })
+    dispatch({ type: DELETE_PRODUCT_REQUEST });
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
-    await axios.delete(`${process.env.REACT_APP_API}/api/product/${id}`, config)
-    dispatch({ type: DELETE_PRODUCT_SUCCESS })
+    };
+    await axios.delete(`/api/product/${id}`, config);
+    dispatch({ type: DELETE_PRODUCT_SUCCESS });
   } catch (error) {
     dispatch({
       type: DELETE_PRODUCT_FAIL,
@@ -156,29 +154,29 @@ export const deleteProduct = (id) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const updateProduct = (product) => async (dispatch, getState) => {
   try {
-    dispatch({ type: UPDATE_PRODUCT_REQUEST })
+    dispatch({ type: UPDATE_PRODUCT_REQUEST });
     const {
       userLogIn: { userInfo },
-    } = getState()
+    } = getState();
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         Authorization: `Bearer ${userInfo.token}`,
       },
-    }
+    };
     await axios.put(
       `${process.env.REACT_APP_API}/api/product/${product.slug}`,
       product,
       config
-    )
-    dispatch({ type: UPDATE_PRODUCT_SUCCESS })
+    );
+    dispatch({ type: UPDATE_PRODUCT_SUCCESS });
   } catch (error) {
     dispatch({
       type: UPDATE_PRODUCT_FAIL,
@@ -186,19 +184,19 @@ export const updateProduct = (product) => async (dispatch, getState) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const detailsProduct = (slug) => async (dispatch) => {
   try {
-    dispatch({ type: DETAILS_PRODUCT_REQUEST })
+    dispatch({ type: DETAILS_PRODUCT_REQUEST });
 
     const { data } = await axios.get(
       `${process.env.REACT_APP_API}/api/product/${slug}`
-    )
+    );
 
-    dispatch({ type: DETAILS_PRODUCT_SUCCESS, payload: data })
+    dispatch({ type: DETAILS_PRODUCT_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: DETAILS_PRODUCT_FAIL,
@@ -206,18 +204,18 @@ export const detailsProduct = (slug) => async (dispatch) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const CountProduct = () => async (dispatch) => {
   try {
-    dispatch({ type: COUNT_PRODUCTS_REQUEST })
+    dispatch({ type: COUNT_PRODUCTS_REQUEST });
 
     const { data } = await axios.get(
       `${process.env.REACT_APP_API}/api/productCount`
-    )
-    dispatch({ type: COUNT_PRODUCTS_SUCCESS, payload: data })
+    );
+    dispatch({ type: COUNT_PRODUCTS_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: COUNT_PRODUCTS_FAIL,
@@ -225,26 +223,26 @@ export const CountProduct = () => async (dispatch) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
 
 export const listProductAdmin = (sort, order, page) => async (dispatch) => {
   try {
-    dispatch({ type: LIST_PRODUCT_ADMIN_REQUEST })
+    dispatch({ type: LIST_PRODUCT_ADMIN_REQUEST });
 
     const config = {
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
-    }
+    };
 
     const { data } = await axios.post(
       `${process.env.REACT_APP_API}/api/productListAdmin`,
       { sort, order, page },
       config
-    )
-    dispatch({ type: LIST_PRODUCT_ADMIN_SUCCESS, payload: data })
+    );
+    dispatch({ type: LIST_PRODUCT_ADMIN_SUCCESS, payload: data });
   } catch (error) {
     dispatch({
       type: LIST_PRODUCT_ADMIN_FAIL,
@@ -252,6 +250,6 @@ export const listProductAdmin = (sort, order, page) => async (dispatch) => {
         error.response && error.response.data.message
           ? error.response.data.message
           : error.message,
-    })
+    });
   }
-}
+};
