@@ -1,20 +1,20 @@
-import React, { useState } from 'react'
-import { Button, Modal, Image, Row, Col, ListGroup } from 'react-bootstrap'
-import { Link } from 'react-router-dom'
-import MultiSelect from 'react-multi-select-component'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import React, { useState } from "react";
+import { Button, Modal, Image, Row, Col, ListGroup } from "react-bootstrap";
+import { Link } from "react-router-dom";
+import { MultiSelect } from "react-multi-select-component";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faHeart,
   faMinus,
   faPlus,
   faTimes,
-} from '@fortawesome/free-solid-svg-icons'
-import { useDispatch } from 'react-redux'
-import { addToCart } from '../actions/cartActions'
-import Skeleton from 'react-loading-skeleton'
-import Message from './Message'
-import { addToWish } from '../actions/userActions'
-import Loader from './Loader'
+} from "@fortawesome/free-solid-svg-icons";
+import { useDispatch } from "react-redux";
+import { addToCart } from "../actions/cartActions";
+import Skeleton from "react-loading-skeleton";
+import Message from "./Message";
+import { addToWish } from "../actions/userActions";
+import Loader from "./Loader";
 
 const FoodModal = ({
   setModalShow,
@@ -25,10 +25,10 @@ const FoodModal = ({
   userInfo,
   loadingAdd,
 }) => {
-  const [counter, setCounter] = useState(1)
-  const [variable, setVariable] = useState('')
-  const [addon, setAddon] = useState([])
-  const dispatch = useDispatch()
+  const [counter, setCounter] = useState(1);
+  const [variable, setVariable] = useState("");
+  const [addon, setAddon] = useState([]);
+  const dispatch = useDispatch();
 
   const handleAddToCart = () => {
     dispatch(
@@ -38,15 +38,15 @@ const FoodModal = ({
         variable,
         addon.length < 1 ? null : addon.map((adn) => adn.value)
       )
-    )
-    setModalShow(false)
-  }
+    );
+    setModalShow(false);
+  };
   function MyVerticallyCenteredModal(props) {
     return (
-      <Modal {...props} size='xl' centered>
+      <Modal {...props} size="xl" centered>
         <Modal.Header
-          className='d-flex justify-content-end pb-0'
-          style={{ backgroundColor: '#FFC94F', border: 'none' }}
+          className="d-flex justify-content-end pb-0"
+          style={{ backgroundColor: "#FFC94F", border: "none" }}
         >
           <Button onClick={props.onHide}>
             <FontAwesomeIcon icon={faTimes} />
@@ -54,13 +54,13 @@ const FoodModal = ({
         </Modal.Header>
         <div
           style={{
-            backgroundColor: '#FFC94F',
-            padding: '50px',
-            borderRadius: '10px',
+            backgroundColor: "#FFC94F",
+            padding: "50px",
+            borderRadius: "10px",
           }}
         >
           <Row>
-            {error && <Message variant='alert'>{error}</Message>}
+            {error && <Message variant="alert">{error}</Message>}
             <Col sm={12} md={6}>
               {loading ? (
                 <Skeleton height={400} count={1} />
@@ -69,20 +69,20 @@ const FoodModal = ({
                   <Image
                     src={product.image && product.image.url}
                     alt={product.title}
-                    className='w-75'
+                    className="w-75"
                   ></Image>
                   {userInfo && (
                     <Button
-                      variant='secondary'
+                      variant="secondary"
                       onClick={() => dispatch(addToWish(product._id))}
                       style={{
-                        position: 'absolute',
-                        top: '10px',
-                        left: '15px',
+                        position: "absolute",
+                        top: "10px",
+                        left: "15px",
                       }}
                     >
                       {loadingAdd ? (
-                        <Loader className='size-sm' />
+                        <Loader className="size-sm" />
                       ) : (
                         <FontAwesomeIcon icon={faHeart} />
                       )}
@@ -97,22 +97,22 @@ const FoodModal = ({
               ) : (
                 <>
                   <ListGroup>
-                    <ListGroup.Item style={{ backgroundColor: 'transparent' }}>
-                      <h2 style={{ color: '#000000', fontWeight: '600' }}>
+                    <ListGroup.Item style={{ backgroundColor: "transparent" }}>
+                      <h2 style={{ color: "#000000", fontWeight: "600" }}>
                         {product.title}
                       </h2>
                     </ListGroup.Item>
-                    <ListGroup.Item style={{ backgroundColor: 'transparent' }}>
+                    <ListGroup.Item style={{ backgroundColor: "transparent" }}>
                       <Row>
                         <Col>Category:</Col>
                         <Col>
                           <Link
                             to={`/category/${product.category?.slug}`}
-                            className='food-card_category'
+                            className="food-card_category"
                             style={{
-                              color: '#34495e',
-                              fontSize: '18px',
-                              cursor: 'pointer',
+                              color: "#34495e",
+                              fontSize: "18px",
+                              cursor: "pointer",
                             }}
                           >
                             {product.category?.name}
@@ -120,7 +120,7 @@ const FoodModal = ({
                         </Col>
                       </Row>
                     </ListGroup.Item>
-                    <ListGroup.Item style={{ backgroundColor: 'transparent' }}>
+                    <ListGroup.Item style={{ backgroundColor: "transparent" }}>
                       <Row>
                         <Col>Description:</Col>
                         <Col>
@@ -128,7 +128,7 @@ const FoodModal = ({
                         </Col>
                       </Row>
                     </ListGroup.Item>
-                    <ListGroup.Item style={{ backgroundColor: 'transparent' }}>
+                    <ListGroup.Item style={{ backgroundColor: "transparent" }}>
                       {product.variable && <h4>Choose from bellow: </h4>}
                       {product.price ? (
                         <h5>Price: {product.price}</h5>
@@ -136,27 +136,27 @@ const FoodModal = ({
                         product.variable.attribute.map((attr) => (
                           <div key={attr._id}>
                             <Row
-                              className='variable-item p-1'
-                              style={{ borderRadius: '10px' }}
+                              className="variable-item p-1"
+                              style={{ borderRadius: "10px" }}
                             >
                               <label htmlFor={attr._id}>
-                                <Col className='d-flex align-items-center'>
+                                <Col className="d-flex align-items-center">
                                   <Image
                                     src={product.image && product.image.url}
                                     alt={product.title}
                                     style={{
-                                      width: '50px',
-                                      margin: '0px 5px',
+                                      width: "50px",
+                                      margin: "0px 5px",
                                     }}
                                   ></Image>
                                   <input
-                                    type='radio'
-                                    name='variable'
+                                    type="radio"
+                                    name="variable"
                                     value={attr._id}
                                     id={attr._id}
                                     onClick={() => setVariable(attr._id)}
                                   />
-                                  <p style={{ margin: '0px' }}>
+                                  <p style={{ margin: "0px" }}>
                                     {attr.name} - ${attr.price}
                                   </p>
                                 </Col>
@@ -166,10 +166,10 @@ const FoodModal = ({
                           </div>
                         ))
                       ) : (
-                        'No Price Given'
+                        "No Price Given"
                       )}
                     </ListGroup.Item>
-                    <ListGroup.Item style={{ backgroundColor: 'transparent' }}>
+                    <ListGroup.Item style={{ backgroundColor: "transparent" }}>
                       {product.addon && product.addon.length > 0 ? (
                         <MultiSelect
                           options={product.addon.map((a) => ({
@@ -178,19 +178,19 @@ const FoodModal = ({
                           }))}
                           value={addon}
                           onChange={setAddon}
-                          labelledBy='Select Addon'
-                          className='product-addons'
+                          labelledBy="Select Addon"
+                          className="product-addons"
                         />
                       ) : (
-                        ''
+                        ""
                       )}
                     </ListGroup.Item>
-                    {product.availability === 'Yes' ? (
+                    {product.availability === "Yes" ? (
                       <ListGroup.Item
-                        style={{ backgroundColor: 'transparent' }}
+                        style={{ backgroundColor: "transparent" }}
                       >
                         <Row>
-                          <Col className='d-flex'>
+                          <Col className="d-flex">
                             <Button
                               onClick={() => setCounter(counter - 1)}
                               disabled={counter < 2}
@@ -199,10 +199,10 @@ const FoodModal = ({
                             </Button>
                             <span
                               style={{
-                                padding: '10px 15px',
-                                backgroundColor: '#fdcb6e',
-                                fontWeight: '600',
-                                fontSize: '18px',
+                                padding: "10px 15px",
+                                backgroundColor: "#fdcb6e",
+                                fontWeight: "600",
+                                fontSize: "18px",
                               }}
                             >
                               {counter}
@@ -216,8 +216,8 @@ const FoodModal = ({
                           </Col>
                           <Col>
                             <Button
-                              style={{ width: '-webkit-fill-available' }}
-                              variant='dark'
+                              style={{ width: "-webkit-fill-available" }}
+                              variant="dark"
                               onClick={handleAddToCart}
                               disabled={product.variable && !variable}
                             >
@@ -227,7 +227,7 @@ const FoodModal = ({
                         </Row>
                       </ListGroup.Item>
                     ) : (
-                      <Message variant='info'>Item Not Available</Message>
+                      <Message variant="info">Item Not Available</Message>
                     )}
                   </ListGroup>
                 </>
@@ -236,7 +236,7 @@ const FoodModal = ({
           </Row>
         </div>
       </Modal>
-    )
+    );
   }
   return (
     <MyVerticallyCenteredModal
@@ -244,7 +244,7 @@ const FoodModal = ({
       onHide={() => setModalShow(false)}
       animation={false}
     />
-  )
-}
+  );
+};
 
-export default FoodModal
+export default FoodModal;
