@@ -1,85 +1,89 @@
-import React, { useEffect, useState } from 'react'
-import FormContainer from '../../components/FormContainer'
-import { Form, Button, Row, Table, Col } from 'react-bootstrap'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { LinkContainer } from 'react-router-bootstrap'
-import { faEdit, faTrash } from '@fortawesome/free-solid-svg-icons'
-import { useDispatch, useSelector } from 'react-redux'
-import Loader from '../../components/Loader'
-import Message from '../../components/Message'
-import { listAttribute } from '../../actions/attributeActions'
+import React, { useEffect, useState } from "react";
+import FormContainer from "../../components/FormContainer";
+import { Form, Button, Row, Table, Col } from "react-bootstrap";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { LinkContainer } from "react-router-bootstrap";
+import { faEdit, faTrash } from "@fortawesome/free-solid-svg-icons";
+import { useDispatch, useSelector } from "react-redux";
+import Loader from "../../components/Loader";
+import Message from "../../components/Message";
+import { listAttribute } from "../../actions/attributeActions";
 import {
   createVariable,
   deleteVariable,
   listVariable,
-} from '../../actions/variableActions'
-import ItemSearch from '../../components/ItemSearch'
-import MultiSelect from 'react-multi-select-component'
+} from "../../actions/variableActions";
+import ItemSearch from "../../components/ItemSearch";
+import { MultiSelect } from "react-multi-select-component";
 
 const VariableCreateScreen = ({ history }) => {
-  const [label, setLabel] = useState('')
-  const [selectedAttribute, setSelectedAttribute] = useState([])
-  const [keyword, setKeyword] = useState('')
-  const dispatch = useDispatch()
+  const [label, setLabel] = useState("");
+  const [selectedAttribute, setSelectedAttribute] = useState([]);
+  const [keyword, setKeyword] = useState("");
+  const dispatch = useDispatch();
   //check logged in user
-  const userLogIn = useSelector((state) => state.userLogIn)
-  const { userInfo } = userLogIn
+  const userLogIn = useSelector((state) => state.userLogIn);
+  const { userInfo } = userLogIn;
 
   //show attribute list
-  const attributeList = useSelector((state) => state.attributeList)
-  const { attributes, error: errorListAttribute } = attributeList
+  const attributeList = useSelector((state) => state.attributeList);
+  const { attributes, error: errorListAttribute } = attributeList;
 
   //Create variable list
-  const variableCreate = useSelector((state) => state.variableCreate)
-  const { loading, success: successCreate, error: errorCreate } = variableCreate
+  const variableCreate = useSelector((state) => state.variableCreate);
+  const {
+    loading,
+    success: successCreate,
+    error: errorCreate,
+  } = variableCreate;
 
   //Show variable list
-  const variableList = useSelector((state) => state.variableList)
-  const { loading: loadingList, variables, error: errorList } = variableList
+  const variableList = useSelector((state) => state.variableList);
+  const { loading: loadingList, variables, error: errorList } = variableList;
 
   //Delete variable
-  const variableDelete = useSelector((state) => state.variableDelete)
+  const variableDelete = useSelector((state) => state.variableDelete);
   const {
     loading: loadingDelete,
     success: successDelete,
     error: errorDelete,
-  } = variableDelete
+  } = variableDelete;
 
   useEffect(() => {
-    if (userInfo && userInfo.role !== 'admin') {
-      history.push('/')
+    if (userInfo && userInfo.role !== "admin") {
+      history.push("/");
     }
-  }, [userInfo, history])
+  }, [userInfo, history]);
 
   useEffect(() => {
-    dispatch(listAttribute())
-    dispatch(listVariable())
+    dispatch(listAttribute());
+    dispatch(listVariable());
     if (successCreate) {
-      setLabel('')
-      setSelectedAttribute([])
+      setLabel("");
+      setSelectedAttribute([]);
     }
-  }, [dispatch, successDelete, successCreate])
+  }, [dispatch, successDelete, successCreate]);
 
   const submitHandler = (e) => {
-    e.preventDefault()
+    e.preventDefault();
     dispatch(
       createVariable({
         name: label,
         attribute: selectedAttribute.map((a) => a.value),
       })
-    )
-  }
+    );
+  };
   const deleteHandler = (id) => {
-    if (window.confirm('Are You Sure?')) {
-      dispatch(deleteVariable(id))
+    if (window.confirm("Are You Sure?")) {
+      dispatch(deleteVariable(id));
     }
-  }
+  };
 
   const searched = (keyword) => (attribute) =>
-    attribute.name.toLowerCase().includes(keyword)
+    attribute.name.toLowerCase().includes(keyword);
   return (
     <>
-      <Row className='align-items-center'>
+      <Row className="align-items-center">
         <Col>
           <h1>Manage Variable</h1>
         </Col>
@@ -90,17 +94,17 @@ const VariableCreateScreen = ({ history }) => {
       </Row>
       <Row>
         <FormContainer>
-          <Form onSubmit={submitHandler} className='my-5'>
-            <Form.Group controlId='label'>
+          <Form onSubmit={submitHandler} className="my-5">
+            <Form.Group controlId="label">
               <Form.Label>Label</Form.Label>
               <Form.Control
-                type='text'
-                placeholder='Enter Label'
+                type="text"
+                placeholder="Enter Label"
                 value={label}
                 onChange={(e) => setLabel(e.target.value)}
               ></Form.Control>
             </Form.Group>
-            <Form.Group controlId='attribute'>
+            <Form.Group controlId="attribute">
               <Form.Label>Attribute</Form.Label>
               <MultiSelect
                 options={attributes.map((a) => ({
@@ -109,23 +113,23 @@ const VariableCreateScreen = ({ history }) => {
                 }))}
                 value={selectedAttribute}
                 onChange={setSelectedAttribute}
-                labelledBy='Select Attributes'
-                className='product-attributes'
+                labelledBy="Select Attributes"
+                className="product-attributes"
               />
             </Form.Group>
-            <Button type='submit' variant='primary' className='my-3'>
+            <Button type="submit" variant="primary" className="my-3">
               Create
             </Button>
           </Form>
         </FormContainer>
       </Row>
       <Row>
-        {errorList && <Message variant='danger'>{errorList}</Message>}
-        {errorDelete && <Message variant='danger'>{errorDelete}</Message>}
+        {errorList && <Message variant="danger">{errorList}</Message>}
+        {errorDelete && <Message variant="danger">{errorDelete}</Message>}
         {errorListAttribute && (
-          <Message variant='danger'>{errorListAttribute}</Message>
+          <Message variant="danger">{errorListAttribute}</Message>
         )}
-        {errorCreate && <Message variant='danger'>{errorCreate}</Message>}
+        {errorCreate && <Message variant="danger">{errorCreate}</Message>}
         {loadingList ? (
           <Loader />
         ) : (
@@ -136,8 +140,8 @@ const VariableCreateScreen = ({ history }) => {
               bordered
               hover
               responsive
-              className='table-sm'
-              variant='dark'
+              className="table-sm"
+              variant="dark"
             >
               <thead>
                 <tr>
@@ -162,13 +166,13 @@ const VariableCreateScreen = ({ history }) => {
                       <LinkContainer
                         to={`/admin/variable/${variable._id}/edit`}
                       >
-                        <Button variant='dark' className='btn-sm'>
+                        <Button variant="dark" className="btn-sm">
                           <FontAwesomeIcon icon={faEdit} />
                         </Button>
                       </LinkContainer>
                       <Button
-                        variant='danger'
-                        className='btn-sm'
+                        variant="danger"
+                        className="btn-sm"
                         onClick={() => deleteHandler(variable._id)}
                       >
                         <FontAwesomeIcon icon={faTrash} />
@@ -182,7 +186,7 @@ const VariableCreateScreen = ({ history }) => {
         )}
       </Row>
     </>
-  )
-}
+  );
+};
 
-export default VariableCreateScreen
+export default VariableCreateScreen;
