@@ -2,8 +2,10 @@ import serviceAccount from "../config/fbServiceAccountKey.json" assert { type: "
 
 var admin = require("firebase-admin");
 
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-});
+// console.log(serviceAccount);
+
+// admin.initializeApp({
+//   credential: admin.credential.cert(serviceAccount),
+// });
 
 export default admin;
